Clean up filenames and pull mechanics.

+++ src/lbry_channel_mirror/sync.py
@@ -5,6 +5,17 @@ import mimetypes
 import time
 import logging
 
+def guess_file_name(stream):
+    def guess_extension(media_type):
+        if media_type == "audio/mpeg":
+            return ".mp3"
+        else:
+            return mimetypes.guess_extension(media_type)
+    title = stream['title']
+    media_type = stream['media_type']
+    return "{name}{ext}".format(
+        name=title, ext=guess_extension(media_type))
+
 def get_channel_id(client, config):
     channel_name = config['channel']
     channel = next(client.resolve({"urls": [channel_name]}))
@@ -25,13 +36,21 @@ def fetch(client, config):
     if not hasattr(config, "claims"):
         config['claims'] = {}
 
+    new_entries = []
     for remote_claim in remote_claims:
-        filename = "{name}{ext}".format(
-            name=remote_claim['name'],
-            ext=mimetypes.guess_extension(remote_claim['value']['stream']['media_type']))
-        config['claims'][remote_claim['claim_id']] = {'file_name': filename}
+        filename = guess_file_name(remote_claim['value']['stream'])
+        claim_id = remote_claim['claim_id']
+        try:
+            # Get existing claim entry:
+            claim = config['claims'][claim_id]
+        except KeyError:
+            # This is a new claim entry:
+            claim = config['claims'][remote_claim['claim_id']] = {'file_name': filename}
+            new_entries.append(claim)
 
-    Config.save(config)
+    if len(new_entries):
+        Config.save(config)
+        logging.info("Added {x} new claims to config: {p}".format(x=len(new_entries), p=config["config_path"]))
 
 def pull(client, config):
     """Download configured files if not existing locally"""
@@ -43,6 +62,7 @@ def pull(client, config):
         for i in claim['items']:
             remote_claims[claim_id] = i
 
+    new_downloads = [] # keep track of each file that didn't exist previously so we can delete later
     downloads = {} # claim_id -> file_list response
     for claim_id, claim in remote_claims.items():
         download_exists = len(next(client.file_list({"claim_id": claim_id}))) > 0
@@ -52,6 +72,9 @@ def pull(client, config):
             if not download_exists:
                 logging.info("Starting Download: {f}".format(f=filename))
             downloads[claim['claim_id']] = next(client.get({"uri": claim['name']}))
+            new_downloads.append(claim_id)
+
+    num_complete = 0
 
     # Wait for downloads to finish, then copy to the final directory:
     while len(downloads):
@@ -64,18 +87,20 @@ def pull(client, config):
                     config['download_directory'], filename)
                 shutil.copy(dl['download_path'], dest_path)
                 # Delete temporary download:
-                if next(client.file_delete({'claim_id': claim_id})):
-                    logging.debug("Deleted temporary download: {f}".format(f=dl['download_path']))
-                else:
-                    logging.warn("Failed to delete temporary downloaded file: {f}".format(
-                        dl['download_path']))
+                if claim_id in new_downloads:
+                    if next(client.file_delete({'claim_id': claim_id})):
+                        logging.debug("Deleted temporary download: {f}".format(f=dl['download_path']))
+                    else:
+                        logging.warn("Failed to delete temporary downloaded file: {f}".format(
+                            dl['download_path']))
 
                 del downloads[claim_id]
                 logging.info("Download Complete: {f}".format(f=filename))
+                num_complete += 1
             else:
                 downloads[claim_id] = next(client.file_list({"claim_id": claim_id}))[0]
                 logging.info("Download Progress: {f} - blobs remaining: {blobs}".format(
                     f=filename, blobs=dl['blobs_remaining']))
                 if dl['blobs_remaining'] > 0:
                     time.sleep(10)
-
+    logging.info("{x} files downloaded.".format(x=num_complete))
